DCOM-93 - Move ReflectionService to Mapping subnamespace to make clearer where it belongs.
