+++ lib/Doctrine/Common/Persistence/Mapping/ReflectionService.php
@@ -17,7 +17,7 @@
  * <http://www.doctrine-project.org>.
  */
 
-namespace Doctrine\Common\Util;
+namespace Doctrine\Common\Persistence\Mapping;
 
 /**
  * Very simple reflection service abstraction.
